+++ package.json
@@ -52,6 +52,7 @@
     "fs": "^0.0.1-security",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.4.0",
+    "mqtt": "^5.10.0",
     "multer": "^1.4.5-lts.1",
     "prettier": "^3.3.2",
     "ts-node-dev": "^2.0.0",
